+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 
 - ✨(backend) allow to leave an item #826
+- ✨(frontend) allow to leave an item #826
 
 ## [v0.22.0] - 2026-09-09